==> R/dissolved_oxygen_percent_saturation/10.1_apply_do_rate_of_change_flag.rmd <==
---
format: 
  html:
    toc: true
    toc-depth: 3
    embed-resources: true
editor_options: 
  chunk_output_type: console
params: 
  county: county
  thresh: thresh
---

```{r, set-up, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, dpi = 600, fig.width = 8)

library(dplyr)
library(DT)
library(ggplot2)
library(here)
library(purrr)
library(qaqcmar)
library(sensorstrings)
library(strings)
library(tgc)
library(tidyr)
library(zoo)

source(here("functions/subchunkify.R"))

theme_set(theme_light())
flag_colours <- c("chartreuse4", "#E6E1BC", "#EDA247", "#DB4325", "grey24")

dt_options <- list(
      dom = 'ft',
      paging = FALSE,
      searching = TRUE,
      pageLength = 500,
      columnDefs = list(list(className = 'dt-center', targets = "_all"))
)

county <- params$county
thresh <- params$thresh

path <- file.path("Y:/coastal_monitoring_program/data_branches/water_quality/processed_data/deployment_data")

path <- paste0(path, "/", county, "/new")

#path <- here("data-raw")

```

```{r, import-data}
#| message: FALSE

dat <- list.files(path, full.names = TRUE, pattern = ".rds") %>% 
  purrr::map(readRDS) %>% 
  list_rbind() %>% 
  select(
    -c(waterbody, lease, latitude, longitude, string_configuration, 
       contains("temperature"), contains("salinity"), contains("uncorrected"),
       contains("sensor_depth_measured")
    )
  ) 

n_hours <- 24

if(!is.numeric(thresh)) stop(paste0("threshold must be numeric, not: ", thresh))

dat_qc <- dat %>% 
  qc_test_rate_of_change(
    rate_of_change_table = data.frame(
      variable = "dissolved_oxygen_percent_saturation",
      stdev_max = thresh
    )
   # keep_sd_cols = TRUE
  )
  # group_by(county, station, deployment_range, sensor_serial_number, variable) %>% 
  # mutate(
  #   # sample interval
  #   int_sample = difftime(timestamp_utc, lag(timestamp_utc), units = "mins"),
  #   int_sample = round(as.numeric(int_sample)),
  # 
  #   # number of samples in n_hours period
  #   n_sample = (60 / int_sample) * n_hours,
  #   n_sample = if_else(is.na(n_sample), 1, n_sample), # first obs
  # 
  #   # backwards rolling sd
  #   sd_roll = rollapplyr(value, width = n_sample, FUN = sd, fill = NA),
  # 
  #   flag_sd = if_else(sd_roll > thresh, 3, 1),
  #   flag_sd = if_else(is.na(flag_sd), 2, flag_sd)
  #   ) %>% 
  # ungroup() 

# summarise flagged observations
qc_summary <- dat_qc %>%
  rename(rate_of_change_flag = rate_of_change_flag_dissolved_oxygen_percent_saturation) %>%
  qc_assign_flag_labels() %>%
  group_by(deployment_range, station, rate_of_change_flag) %>%
  summarise(flag_count = n()) %>%
  mutate(flag_percent = round((100 * flag_count / sum(flag_count)), digits = 3)) %>%
  ungroup() %>%
  select(-flag_count) %>%
  pivot_wider(names_from = "rate_of_change_flag", values_from = "flag_percent") %>%
  arrange(station, deployment_range)
```

# `r county` QC Flags: Dissolved Oxygen, Rate of Change

`r Sys.Date()`


# Summary Table

Percent of observations in each deployment assigned with each qc flag. 

```{r}
datatable(qc_summary, options = dt_options, rownames = FALSE)
```

# Figures

```{r, body} 
#| results: 'asis'
#| message: FALSE

#if("Suspect/Of Interest" %in% colnames(qc_summary)) {
  
  depl_id <- qc_summary %>%
  #  filter(!is.na(`Suspect/Of Interest`)) %>% 
    distinct(deployment_range, station) 
  
  stations <- unique(depl_id$station)
  
  for(i in seq_along(stations)){
    
    station_i <- stations[i]
    
    deployments <- depl_id %>% filter(station == station_i)
    deployments <- deployments$deployment_range
    
    cat('\n##', station_i, '\n')
    
    cat('\n')
    
    # plot for each deployment
    for(j in 1:length(deployments)){
      
      depl_j <- deployments[j]
      
      # subset data to station of interest
      dat_j <- dat_qc %>%
        filter(station == station_i, deployment_range == depl_j) %>%
        qc_pivot_longer(qc_tests = "rate_of_change") %>% 
        ss_convert_depth_to_ordered_factor()
      
      # height of figure j
      n_sensor <- length(unique(dat_j$sensor_serial_number))
      
      if(n_sensor == 1) h = 2.75
      if(n_sensor > 1) h = 2.75 * n_sensor
      
      
      p <- qc_plot_flags(dat_j)
      p <- p$dissolved_oxygen_percent_saturation$rate_of_change

        # ggplot(dat_j, aes(timestamp_utc, value, col = factor(flag_sd))) +
        # geom_point(size = 0.5) +
        # scale_colour_manual("Flag Value", values = flag_colours, drop = FALSE) +
        # facet_wrap(~sensor_serial_number, ncol = 1, scales = "free_y")

      cat('\n###', depl_j, '\n')
      
      subchunkify(p, fig_height = h, fig_width = 8.5)
      
      #cat(paste("Figure ", k, ": ", station.j, fig.caption,  sep = ""), '\n')
      
      cat('\n')
    }
  }


```
